# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: StellarRoute API
  version: 0.1.0
  description: |
    REST API for DEX aggregation on the Stellar Network.

    StellarRoute aggregates liquidity from Stellar's SDEX (Stellar Distributed
    Exchange) and Soroban AMM pools to find the best available price for any
    token pair.

    ## Asset identifiers

    Assets are encoded in path segments using one of three formats:

    | Format | Example | Meaning |
    |--------|---------|---------|
    | `native` | `native` | XLM (Stellar's native asset) |
    | `CODE` | `USDC` | Issued asset — code only (omits issuer) |
    | `CODE:ISSUER` | `USDC:GA5ZSEJ...` | Issued asset — fully qualified |

    ## Error handling

    All errors return a JSON body with `error` (machine-readable code) and
    `message` (human-readable description).  See the `ErrorResponse` schema
    and the **Error codes** section below for the full list.

    ## Rate limiting

    Endpoints are rate-limited per IP address.  When the limit is exceeded the
    API returns **429 Too Many Requests** with `X-RateLimit-*` headers.

    ## Request correlation

    Clients may optionally send `X-Request-ID` with any request. The API will
    reuse that value, or generate a UUID when the header is omitted, and echo
    the final request ID back in the `X-Request-ID` response header.

    ## Versioning and deprecation

    `/api/v1/*` is in deprecation mode. Responses from v1 routes include
...
    notice period and migration guidance.
...
servers:
...
paths:
...
  /api/v1/orderbook/{base}/{quote}:
    get:
      operationId: get_orderbook
      summary: Get orderbook
      description: |
        Returns bids and asks for the given base/quote asset pair.

        **Bids** are orders to buy the base asset with the quote asset
        (sorted highest price first).  **Asks** are orders to sell the base
        asset for the quote asset (sorted lowest price first).

        Results are cached for **10 seconds**.
      tags:
        - trading
      parameters:
        - $ref: "#/components/parameters/BaseAsset"
        - $ref: "#/components/parameters/QuoteAsset"
      responses:
        "200":
          description: Orderbook data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderbookResponse"
              example:
                base_asset:
                  asset_type: native
                  asset_code: null
                  asset_issuer: null
                quote_asset:
                  asset_type: credit_alphanum4
                  asset_code: USDC
                  asset_issuer: GA5ZSEJYB37JRC5AVCIA5MOP4RHTM335X2KGX3IHOJAPP5RE34K4KZVN
                bids:
                  - price: "0.1050000"
                    amount: "500.0000000"
                    total: "52.5000000"
                asks:
                  - price: "0.1060000"
                    amount: "300.0000000"
                    total: "31.8000000"
                summary:
                  bid: "0.1050000"
                  ask: "0.1060000"
                  spread_bps: 95
                  midpoint: "0.1055000"
                timestamp: 1740312000
        "400":
          description: Invalid asset identifier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                v: 1
                timestamp: 1740312000000
                request_id: req_01hyxk6bzv4n9p8m8j1f4c0a2r
                data:
                  error: invalid_asset
                  message: "Invalid base asset: unknown asset format"
        "404":
          description: Asset pair not found in the orderbook
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                v: 1
                timestamp: 1740312000000
                request_id: req_01hyxk6bzv4n9p8m8j1f4c0a2r
                data:
                  error: not_found
                  message: Asset not found in orderbook
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                v: 1
                timestamp: 1740312000000
                request_id: req_01hyxk6bzv4n9p8m8j1f4c0a2r
                data:
                  error: internal_error
                  message: An internal error occurred
        "429":
          $ref: "#/components/responses/RateLimited"
...
  /api/v1/route/{base}/{quote}:
    get:
      operationId: get_route
      summary: Get trading route
      deprecated: true
      description: |
        Returns only the optimal execution path and routing steps for trading
        `amount` of the base asset for the quote asset.

        Unlike the quote endpoint, this returns a simplified response without
        detailed pricing breakdown or venue evaluations.

        This legacy route is deprecated in favor of `/api/v1/routes/{base}/{quote}`.
        Clients should watch the `Deprecation`, `Sunset`, and `Link` headers and
        migrate using the guidance in `docs/api/versioning.md`.
      tags:
        - trading
      parameters:
        - $ref: "#/components/parameters/BaseAsset"
        - $ref: "#/components/parameters/QuoteAsset"
        - name: amount
          in: query
          description: Amount of the base asset to trade (default `1`)
          required: false
          schema:
            type: string
            example: "100"
        - name: slippage_bps
          in: query
          description: Slippage tolerance in basis points (e.g. `50` = 0.50%). Max `10000` (100%).
          required: false
          schema:
            type: integer
            default: 50
            example: 100
        - name: quote_type
          in: query
          description: |
            Direction of the quote:
            - `sell` — how much quote asset you receive when **selling** `amount` of the base asset
            - `buy`  — how much base asset you must spend to **buy** `amount` of the quote asset
          required: false
          schema:
            type: string
            enum: [sell, buy]
            default: sell
      responses:
        "200":
          description: Optimal trading route
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteResponse"
        "400":
          description: Invalid parameters (bad asset format or invalid slippage)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No trading route found for this pair
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          $ref: "#/components/responses/RateLimited"
components:
  parameters:
...
    BaseAsset:
      name: base
      in: path
      required: true
      description: |
        Base asset identifier. One of:
        - `native` — XLM
        - `CODE` — issued asset code only (e.g. `USDC`)
        - `CODE:ISSUER` — fully-qualified issued asset (e.g. `USDC:GA5Z...`)
      schema:
        type: string
        example: native
    QuoteAsset:
      name: quote
      in: path
      required: true
      description: |
        Quote asset identifier. One of:
        - `native` — XLM
        - `CODE` — issued asset code only (e.g. `USDC`)
        - `CODE:ISSUER` — fully-qualified issued asset (e.g. `USDC:GA5Z...`)
      schema:
        type: string
        example: USDC
...
  responses:
    RateLimited:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          description: Maximum number of requests allowed in the current window
          schema:
            type: integer
        X-RateLimit-Remaining:
          description: Number of requests remaining in the current window
          schema:
            type: integer
        X-RateLimit-Reset:
          description: Unix timestamp when the rate limit window resets
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: rate_limit_exceeded
            message: Too many requests — please slow down
...
  schemas:
...
    AssetInfo:
      type: object
      required: [asset_type]
      description: Stellar asset descriptor
      properties:
        asset_type:
          type: string
          description: Stellar asset type
          enum: [native, credit_alphanum4, credit_alphanum12]
          example: credit_alphanum4
        asset_code:
          type: string
          nullable: true
          description: Asset code (null for native XLM)
          example: USDC
        asset_issuer:
          type: string
          nullable: true
          description: G-address of the issuing account (null for native XLM)
          example: GA5ZSEJYB37JRC5AVCIA5MOP4RHTM335X2KGX3IHOJAPP5RE34K4KZVN
...
    OrderbookLevel:
      type: object
      required: [price, amount, total]
      description: A single price level in the orderbook
      properties:
        price:
          type: string
          description: Price as a decimal string (7 decimal places)
          example: "0.1050000"
        amount:
          type: string
          description: Available amount at this price level
          example: "500.0000000"
        total:
          type: string
          description: Total value at this price level (`price × amount`)
          example: "52.5000000"
    OrderbookResponse:
      type: object
      required: [base_asset, quote_asset, bids, asks, timestamp]
      description: Full orderbook snapshot for a trading pair
      properties:
        base_asset:
          $ref: "#/components/schemas/AssetInfo"
        quote_asset:
          $ref: "#/components/schemas/AssetInfo"
        bids:
          type: array
          description: Buy orders sorted highest price first
          items:
            $ref: "#/components/schemas/OrderbookLevel"
        asks:
          type: array
          description: Sell orders sorted lowest price first
          items:
            $ref: "#/components/schemas/OrderbookLevel"
        timestamp:
          type: integer
          format: int64
          description: Unix timestamp of the snapshot
          example: 1740312000
...
    PathStep:
      type: object
      required: [from_asset, to_asset, price, source]
      description: A single hop in the optimal execution path
      properties:
        from_asset:
          $ref: "#/components/schemas/AssetInfo"
        to_asset:
          $ref: "#/components/schemas/AssetInfo"
        price:
          type: string
          description: Exchange rate for this hop
          example: "0.1055000"
        source:
          type: string
          description: |
            Liquidity source for this hop:
            - `sdex` — Stellar Distributed Exchange
            - `amm:<pool_address>` — Soroban AMM pool
          example: sdex
...
    RouteResponse:
      type: object
      required: [base_asset, quote_asset, amount, path, slippage_bps, timestamp]
      description: Optimal trading route without pricing details
      properties:
        base_asset:
          $ref: "#/components/schemas/AssetInfo"
        quote_asset:
          $ref: "#/components/schemas/AssetInfo"
        amount:
          type: string
          description: Amount of base asset being traded
          example: "100.0000000"
        path:
          type: array
          description: List of execution steps
          items:
            $ref: "#/components/schemas/PathStep"
        slippage_bps:
          type: integer
          description: Slippage tolerance in basis points
          example: 50
        timestamp:
          type: integer
          format: int64
          description: Unix timestamp of the route calculation
          example: 1740312000
    ErrorResponse:
      type: object
      required: [error, message]
      description: Standard API error format
      properties:
        error:
          type: string
          description: Machine-readable error code
          example: validation_error
        message:
          type: string
          description: Human-readable error description
          example: "Amount must be greater than zero"
        details:
          type: object
          nullable: true
          description: Optional structured context about the error
          example:
            parameter: slippage_bps
            min: 0
            max: 10000

  # Authentication is planned for a future release.
  # securitySchemes:
  #   bearerAuth:
  #     type: http
  #     scheme: bearer
  #     bearerFormat: JWT
  #     description: |
  #       API key authentication (Phase 2).
  #       Pass a JWT obtained from the /auth/token endpoint.